worked on routing and responsivesness

+++ src/app/sign-up/sign-up.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-sign-up',
@@ -6,16 +7,36 @@ import { Component } from '@angular/core';
   styleUrls: ['./sign-up.component.css']
 })
 export class SignUpComponent {
+  constructor (public route: Router) {}
   public full_name = "";
   public phone_no = "";
   public email = "";
   public password  = "";
   public usersArray:any = [];
+  public message = false;
+
+  ngOnInit () {
+    let getUsers:any = localStorage.getItem("taskUsers");
+    getUsers = JSON.parse(getUsers);
+    console.log(getUsers, "getUsers")
+    if (getUsers) {
+      this.usersArray = getUsers
+    } else {
+      this.usersArray = [];
+    }
+    console.log(this.usersArray, "usersArray")
+  }
 
 
   signup () {
-    let obj = {full_name:this.full_name, phone_no:this.phone_no, email:this.email, password:this.password};
-    this.usersArray.push(obj);
-    localStorage.setItem("taskUsers",JSON.stringify(this.usersArray));
+    let checkExist = this.usersArray.find((each:any) => each.email == this.email);
+    if (checkExist) {
+      this.message = true
+    }else{
+      let obj = {full_name:this.full_name, phone_no:this.phone_no, email:this.email, password:this.password};
+      this.usersArray.push(obj);
+      localStorage.setItem("taskUsers",JSON.stringify(this.usersArray));
+      this.route.navigate(["/addtask"]);
+    }
   }
 }
